Move model class Station to domain module
