--- domainfeature1/impl/src/main/java/dev/xxxxx/domainfeature1/impl/MyClass.kt
@@ -1,5 +1,0 @@
-package dev.xxxxx.domainfeature1.impl
-
-class MyClass {
-    /* no-op */
-}
