--- domainfeature1/public/src/main/java/dev/xxxxx/domainfeature1/MyClass.kt
@@ -1,5 +1,0 @@
-package dev.xxxxx.domainfeature1
-
-class MyClass {
-    /* no-op */
-}
