+++ domainfeature1/public/src/main/java/dev/xxxxx/domainfeature1/Station.kt
@@ -1,11 +1,9 @@
-package dev.xxxxx.feature1
+package dev.xxxxx.domainfeature1
 
 data class Station(
         val id: String,
         val label: String,
         val address: String,
-        val cityName: String,
-        val provinceName: String,
         val gas95E5Price: String,
         val gas98E5Price: String,
         val aGasPrice: String,
