--- feature1/src/main/java/dev/xxxxx/feature1/MyClass.kt
@@ -1,5 +1,0 @@
-package dev.xxxxx.feature1
-
-class MyClass {
-    /* no-op */
-}
